flag consolidation and proxy file #issue1001
